ath11k: Add support for QCA6x90 in MSM platform
QCA6x90 in MSM soc is not connected over M.2 connector and it
needs few regulators, pin ctrl to be turned on for PCI
enumeration and needs these resources to be turned off to
reduce leakage current during driver removal.
There are 2 different types of QCA6x90 module available which
needs different supplies to be turned on and can be attached on
sm8150/sm8250 MSM SOC.

To support these SKU, add platform driver to vote for regulators and
pin ctrl resources.

Signed-off-by: Govind Singh <govinds@codeaurora.org>
Signed-off-by: Govind Singh <govind.sk85@gmail.com>